requirements.txt
httpx
markdown-it-py[linkify]
